requirements.txt
google-adk==1.19.0
feedparser==6.0.12
python-dotenv==1.2.1